endpoint

+++ models/Node.h
@@ -42,7 +42,7 @@ class Node
 
         void updateRectangle() {
             if (leaf) {
-                for (size_t i = 0; i < count; i += 1) {
+                for (int i = 0; i < count; i += 1) {
                     if (polygons[i]->min.x < rectangle->min.x) {
                         rectangle->min.x = polygons[i]->min.x;
                     }
@@ -57,7 +57,7 @@ class Node
                     }
                 }
             } else {
-                for (size_t i = 0; i < count; i += 1) {
+                for (int i = 0; i < count; i += 1) {
                     if (children[i]->rectangle->min.x < rectangle->min.x) {
                         rectangle->min.x = children[i]->rectangle->min.x;
                     }
@@ -93,8 +93,8 @@ class Node
 
         void getFartherPolygons(int & a, int & b) {
             float maxDistance = std::numeric_limits<T>::min();
-            for (size_t i = 0; i < count; i += 1) {
-                for (size_t j = i + 1; j < count; j += 1) {
+            for (int i = 0; i < count; i += 1) {
+                for (int j = i + 1; j < count; j += 1) {
                     float newDistance = polygons[i]->getDistance(*polygons[j]);
                     if (newDistance > maxDistance) {
                         maxDistance = newDistance;
